+++ lib/src/contracts_generator_config.dart
@@ -11,6 +11,7 @@ class ContractsGeneratorConfig {
     this.name = 'contracts',
     RegExp? include,
     Directory? output,
+    this.directives = '',
     this.extra = '',
   })  : include = include ?? RegExp(''),
         output = output ?? Directory.current;
@@ -24,6 +25,7 @@ class ContractsGeneratorConfig {
     final dynamic name = config['name'];
     final dynamic include = config['include'];
     final dynamic output = config['output'];
+    final dynamic directives = config['directives'];
     final dynamic extra = config['extra'];
 
     if (name is! String) {
@@ -35,6 +37,9 @@ class ContractsGeneratorConfig {
     if (output is! String) {
       throw ArgumentError('`output` field has to be a string');
     }
+    if (directives is! String) {
+      throw ArgumentError('`directives` field has to be a string');
+    }
     if (extra is! String) {
       throw ArgumentError('`extra` field has to be a string');
     }
@@ -44,6 +49,7 @@ class ContractsGeneratorConfig {
       name: name,
       include: RegExp(include),
       output: Directory(output),
+      directives: directives,
       extra: extra,
     );
   }
@@ -60,7 +66,10 @@ class ContractsGeneratorConfig {
   /// Output directory
   final Directory output;
 
-  /// Text to be added in generated contracts file between imports and statements
+  /// Directives (`export`, `import`, `part` etc.) to be added to the generated contracts file
+  final String directives;
+
+  /// Code to be added to the generated contracts file after all directives
   final String extra;
 
   static const defaultYamlConfig = '''
@@ -86,7 +95,9 @@ name: contracts
 include: .*
 # Output directory
 output: lib/data
-# Text to be added in generated contracts file between imports and statements
+# Directives (`export`, `import`, `part` etc.) to be added to the generated contracts file
+directives: ""
+# Code to be added to the generated contracts file after all directives
 extra: ""
 ''';