+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Unreleased
 
-- **BREAKING**: Add support for `project`, `file`, and `path` input methods in config file
+- **BREAKING**: Add support for `project` and `path` input methods in config file
+- Add directives config option
 - Fix incorrect field renaming
 - Stopped generating redundant dart doc from constructor parameters
 - Annotate generic classes with `genericArgumentFactories: true`